🛠️ Refactor: Directory Layout and C Header Architecture
- **Context:** The codebase previously contained all logic in a flat root directory and relied heavily on the C anti-pattern of `#include "file.c"`. Additionally, there was no centralized error handling system.
- **Extraction and Naming:** The structure was reorganized into domain-specific modules (`src/core`, `src/utils`, `src/cli`, `tests`). We created proper `.h` header files for all corresponding `.c` implementations to enforce proper encapsulation and separate compilation units according to C standards.
- **Centralized Error Reporting:** Created `tfs_report_error` inside `src/utils/log.c/h` to act as the primary funnel for unexpected or terminal errors as requested in global constraints, and migrated sample calls in `cli.c`.
- **Operational Memory:** Generated `AGENTS.md` containing the project layout map, inclusion rules, and compilation instructions to guide future agents.
- **Testing:** Adapted `build.sh` and verified `cli.c` main interactions.

### Assumptions
- The `main.c` file present in the root previously was effectively a test/demo, as `cli.c` contains the actual `main` entrypoint for the application. Thus, `main.c` was relocated to `tests/`.
- No new linters or formats were introduced, respecting the requirement to not install dependencies manually without `mise` tooling configs.

### Alternatives Not Chosen
- Continuing to allow `#include "file.c"`: This would persist broken architectural practices and lead to multiple definition errors if more than one `.c` file included the same module.

### How To Pivot
- If the directory names are not ideal, move the folders in `src/` and adjust `#include "../folder/"` paths and `build.sh` accordingly.

### Next Knobs
- `src/utils/log.c`: Add real Sentry integration or richer contextual log fields to `tfs_report_error`.
- Replace all remaining `tfs_log_err` with `tfs_report_error` throughout the codebase.
