+++ src/core/defs.h
