+++ src/core/tree_prettyprint.h
@@ -1,0 +1,9 @@
+#ifndef _TFS_TREE_PRETTYPRINT_H
+#define _TFS_TREE_PRETTYPRINT_H
+
+#include "defs.h"
+
+void tfs_members__print(struct tfs_members_t *mems, int level);
+void tfs_node__print(struct tfs_node_t *node, int level);
+
+#endif
