+++ src/utils/log.h
@@ -1,10 +1,8 @@
-/*
- * Utilitários para mostra de relatórios mais coloridos
- */
+#ifndef _TFS_LOG_H
+#define _TFS_LOG_H
 
-#ifndef _TFS_LOG
-#define _TFS_LOG
 #include <stdio.h>
+
 #define _tfs_log_prefix(boldness, color, prefix) \
     printf("\033[%i;%im%c: \033[0m", boldness, color, prefix);
 
@@ -20,5 +18,6 @@
     _tfs_log_prefix(0, 34, 'I'); \
     printf(__VA_ARGS__); 
 
+void tfs_report_error(const char *msg);
 
 #endif
