+++ src/utils/string_stack.h
